more clean up in app
